Update app good table

+++ pkg/db/ent/appgood.go
@@ -31,11 +31,11 @@ type AppGood struct {
 	// InvitationOnly holds the value of the "invitation_only" field.
 	InvitationOnly bool `json:"invitation_only,omitempty"`
 	// CreateAt holds the value of the "create_at" field.
-	CreateAt int64 `json:"create_at,omitempty"`
+	CreateAt uint32 `json:"create_at,omitempty"`
 	// UpdateAt holds the value of the "update_at" field.
-	UpdateAt int64 `json:"update_at,omitempty"`
+	UpdateAt uint32 `json:"update_at,omitempty"`
 	// DeleteAt holds the value of the "delete_at" field.
-	DeleteAt int64 `json:"delete_at,omitempty"`
+	DeleteAt uint32 `json:"delete_at,omitempty"`
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
@@ -118,19 +118,19 @@ func (ag *AppGood) assignValues(columns []string, values []interface{}) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field create_at", values[i])
 			} else if value.Valid {
-				ag.CreateAt = value.Int64
+				ag.CreateAt = uint32(value.Int64)
 			}
 		case appgood.FieldUpdateAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field update_at", values[i])
 			} else if value.Valid {
-				ag.UpdateAt = value.Int64
+				ag.UpdateAt = uint32(value.Int64)
 			}
 		case appgood.FieldDeleteAt:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field delete_at", values[i])
 			} else if value.Valid {
-				ag.DeleteAt = value.Int64
+				ag.DeleteAt = uint32(value.Int64)
 			}
 		}
 	}
